cleaned mvp and renamed it fixed tests
